requirements.txt
alabaster==0.7.13
asgiref==3.6.0
Babel==2.12.1
boto3==1.26.101
botocore==1.29.101
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
cryptography==40.0.1
dataclasses==0.6
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-filter==23.1
django-tables2==2.5.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.19
et-xmlfile==1.1.0
Faker==18.3.1
filelock==3.10.7
firebase-admin==6.0.0
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==20.1.0
httplib2==0.22.0
identify==2.5.22
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
msgpack==1.0.5
mysqlclient==2.1.1
Naked==0.1.32
nodeenv==1.7.0
openpyxl==3.1.2
packaging==23.0
paytmchecksum==1.7.0
Pillow==9.4.0
platformdirs==3.2.0
pre-commit==3.2.1
proto-plus==1.22.2
protobuf==4.22.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
pyfcm==1.5.4
Pygments==2.14.0
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
razorpay==1.3.0
requests==2.28.2
rest-condition==1.0.3
rsa==4.9
s3transfer==0.6.0
shellescape==3.8.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
tablib==3.4.0
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
xlrd==1.2.0
XlsxWriter==3.0.9